requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
authlib==1.3.0
httpx==0.25.1
chromadb
langchain==0.0.350
transformers==4.36.0
torch==2.2.0
google-auth-oauthlib==1.1.0
google-auth==2.23.4
google-api-python-client==2.108.0
notion-client==2.1.0
python-jose==3.3.0
python-multipart==0.0.6 
requests 